Reverse relation to denote a task and its dependencies.
Previously, the child task would get an ID assigned and the parent task
would refer to it. Now it's the other way around: the parent task gets
an ID and the dependency refers to it's parent using the 'p' tag. The
'dep' tag is no longer used.

The problem with the former situation is that when you complete a
dependency, its ID would disappear from the todo.txt. When adding a new
dependency elsewhere, the task would get assigned the ID you've just
thrown away. And there are still parent tasks lingering around to this
task.

The current scheme works better, assuming that you finish (or delete)
subtasks before you finish the parent task.