Add 'dot' subcommand to 'dep' for making Dot graphs of dependencies.
The dot subcommand takes a task number or a text filter.

When given a task number, all direct and indirect dependencies of the
given task will be represented in the Dot graph.

When given a text, all tasks containing that text will be used and all
dependencies between those will be visible. This can be used to filter
for a certain project or context.

Example:

    todo.sh dep dot +SecretProject

The output still looks a bit crude, might be tuned later on.